Backend/requirements.txt
Django==4.2.23
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
python-decouple==3.8
celery==5.3.4
redis==5.0.1